+++ Dockerfile.pubsub
@@ -1,0 +1,13 @@
+FROM python:3.12-slim
+
+# Keeps Python from generating .pyc files in the container
+ENV PYTHONDONTWRITEBYTECODE=1
+# Turns off buffering for easier container logging
+ENV PYTHONUNBUFFERED=1
+
+WORKDIR /app
+COPY ./pubsub.py .
+COPY ./requirements.txt .
+RUN python -m pip install -r requirements.txt
+
+CMD ["python", "pubsub.py"]
